+++ ntex/src/framed/mod.rs
@@ -1,11 +1,9 @@
-mod dispatcher;
-mod error;
 mod handshake;
 mod service;
 
-pub use self::error::ServiceError;
 pub use self::handshake::{Handshake, HandshakeResult};
 pub use self::service::{Builder, FactoryBuilder};
+pub use crate::util::framed::DispatcherError as ServiceError;
 
 #[doc(hidden)]
 pub type Connect<T, U> = Handshake<T, U>;