+++ ntex/src/ws/mod.rs
@@ -8,13 +8,11 @@ use std::io;
 use derive_more::{Display, From};
 
 mod codec;
-mod dispatcher;
 mod frame;
 mod mask;
 mod proto;
 
 pub use self::codec::{Codec, Frame, Item, Message};
-pub use self::dispatcher::Dispatcher;
 pub use self::frame::Parser;
 pub use self::proto::{hash_key, CloseCode, CloseReason, OpCode};